+++ CHANGELOG.md
@@ -6,10 +6,12 @@ Version 1.10-SNAPSHOT
 
 CHANGED:
 
-- update of jobj-core dependency version to 5.1
-- update of gradle-plugin com.github.ben-manes.versions to version 0.40.0
+- update of gradle-plugin com.github.ben-manes.versions to version 0.41.0
+- update of jobj-core dependency version to 5.2
 - remove of jobj-copy dependency
 - remove of jobj-merge dependency
+- update of bean-mapper-api dependency version to 1.2
+- update of generic-mapper dependency version to 2.6
 
 Version 1.9
 -------------